+++ src/js/easymde.js
@@ -12,7 +12,7 @@ require('codemirror/addon/search/searchcursor.js');
 require('codemirror/mode/gfm/gfm.js');
 require('codemirror/mode/xml/xml.js');
 var CodeMirrorSpellChecker = require('codemirror-spell-checker');
-var marked = require('marked/lib/marked');
+var marked = require('marked').marked;
 
 
 // Some variables
@@ -1986,7 +1986,7 @@ EasyMDE.prototype.markdown = function (text) {
         marked.setOptions(markedOptions);
 
         // Convert the markdown to HTML
-        var htmlText = marked(text);
+        var htmlText = marked.parse(text);
 
         // Sanitize HTML
         if (this.options.renderingConfig && typeof this.options.renderingConfig.sanitizerFunction === 'function') {