+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Fixed
 - Incorrect initial line and column count in status bar.
+- Security issue in `marked` dependency.
 
 ## [2.16.0] - 2022-01-11
 ### Added